FastAPI/requirements.txt
llama-index
python-dotenv
qdrant_client
llama-index-llms-groq
llama-index-embeddings-huggingface
elevenlabs
selenium
webdriver_manager
SpeechRecognition
whisper
pyaudio
langchain
pydub
langchain-groq
simpleaudio
llama-index-embeddings-ollama
fastapi
uvicorn